+++ include/LayoutMirror.hpp
@@ -4,7 +4,6 @@
 #include <deque>
 #include <string>
 #include <unordered_map>
-#include <unordered_set>
 #include <vector>
 
 class LayoutMirror final {
@@ -19,7 +18,7 @@ public:
 
     void prepareFor(PlayLayer* real, GJGameLevel* level);
     void observeObject(PlayLayer* real, GameObject* object);
-    void observeVisibility(GameObject* object, bool visible);
+    void observeOpacity(GameObject* object, unsigned char opacity);
     void finishFor(PlayLayer* real);
     void destroyFor(PlayLayer* real);
     void renderPlayerView(cocos2d::CCDirector* director, PlayLayer* real);
@@ -39,6 +38,8 @@ private:
         GameObject* object = nullptr;
         bool keep = false;
         bool forceHidden = false;
+        unsigned char layoutOpacity = 255;
+        std::uint64_t touchedSerial = 0;
     };
 
     struct SavedVisualState {
@@ -46,7 +47,9 @@ private:
         cocos2d::CCSprite* sprite = nullptr;
         cocos2d::CCParticleSystemQuad* particle = nullptr;
         cocos2d::ccColor3B color {255, 255, 255};
+        unsigned char opacity = 255;
         bool restoreColor = false;
+        bool restoreOpacity = false;
         bool restoreVisible = false;
         bool visible = false;
     };
@@ -79,7 +82,8 @@ private:
     void endLayoutPass();
     void restoreLayoutOverrides();
     void registerRenderNodes(LayoutEntry const& entry);
-    void applyBatchedOverrides();
+    void applyCameraOverrides(PlayLayer* real);
+    void touchCameraEntry(LayoutEntry& entry);
     void applyScenePalette(PlayLayer* real);
     void saveAndColorSceneSprite(cocos2d::CCSprite* sprite, cocos2d::ccColor3B color);
 
@@ -87,12 +91,12 @@ private:
     std::string m_modifiedString;
     std::vector<LayoutEntry> m_entries;
     std::unordered_map<GameObject*, std::size_t> m_entryIndex;
-    std::unordered_set<std::size_t> m_visibleEntries;
     std::unordered_map<cocos2d::CCNode*, RenderNodeEntry> m_renderNodes;
     std::unordered_map<int, std::deque<PendingRecord>> m_pendingByObjectID;
     std::unordered_map<int, cocos2d::ccColor3B> m_layoutPalette;
     std::vector<SavedVisualState> m_savedStates;
     std::vector<SavedSceneSprite> m_savedSceneSprites;
+    std::uint64_t m_frameSerial = 0;
     std::size_t m_originalRecordCount = 0;
     std::size_t m_transformedRecordCount = 0;
     std::size_t m_classifiedKeepCount = 0;
@@ -103,7 +107,9 @@ private:
     std::size_t m_frameMappedNodeCount = 0;
     std::size_t m_frameStyledNodeCount = 0;
     std::size_t m_frameSuppressedNodeCount = 0;
-    std::size_t m_frameActiveObjectCount = 0;
+    std::size_t m_frameCandidateObjectCount = 0;
+    std::size_t m_frameRetainedCandidateCount = 0;
+    std::size_t m_frameForcedVisibleCount = 0;
     std::size_t m_frameBatchedMutationCount = 0;
     bool m_reportedRenderCoverage = false;
     bool m_renderMapReady = false;